+++ apps/web/app/(minimal)/new/newEventProgressStages.tsx
